% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-datalad.R
\name{.datalad_action}
\alias{.datalad_action}
\title{Determine DataLad Action for Directory}
\usage{
.datalad_action(dest_dir)
}
\arguments{
\item{dest_dir}{Path to the destination directory.}
}
\value{
Character string: "clone" if directory doesn't exist or is empty,
"update" if directory is already a DataLad dataset. Aborts if directory
exists but is not a DataLad dataset.
}
\description{
Checks whether a destination directory needs to be cloned or can be updated.
}
\keyword{internal}
