% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-management.R
\name{.empty_cache_tibble}
\alias{.empty_cache_tibble}
\title{Create Empty Cache Tibble}
\usage{
.empty_cache_tibble()
}
\value{
An empty tibble with cache list columns.
}
\description{
Returns an empty tibble with the correct column structure for cache list.
}
\keyword{internal}
