% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctor.R
\name{.get_aws_version}
\alias{.get_aws_version}
\title{Get AWS CLI Version}
\usage{
.get_aws_version()
}
\value{
Character string with version, or NA_character_ on failure.
}
\description{
Runs \code{aws --version} and parses the output to extract version string.
}
\keyword{internal}
