% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctor.R
\name{.get_datalad_version}
\alias{.get_datalad_version}
\title{Get DataLad Version}
\usage{
.get_datalad_version()
}
\value{
Character string with version, or NA_character_ on failure.
}
\description{
Runs \code{datalad --version} and parses the output to extract version string.
}
\keyword{internal}
