% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery-github.R
\name{.github_after}
\alias{.github_after}
\title{Calculate Retry Delay for Rate Limited Response}
\usage{
.github_after(resp)
}
\arguments{
\item{resp}{An httr2 response object.}
}
\value{
Number of seconds to wait before retry (minimum 0).
}
\description{
Calculates how many seconds to wait before retrying a
rate-limited GitHub API request.
}
\keyword{internal}
