% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject-filter.R
\name{.normalize_subject_ids}
\alias{.normalize_subject_ids}
\title{Normalize Subject IDs}
\usage{
.normalize_subject_ids(ids)
}
\arguments{
\item{ids}{Character vector of subject IDs.}
}
\value{
Character vector with all IDs prefixed with "sub-".
}
\description{
Vectorized wrapper for \code{\link[=.normalize_subject_id]{.normalize_subject_id()}}.
}
\keyword{internal}
