% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-response.R
\name{.parse_datasets_response}
\alias{.parse_datasets_response}
\title{Parse Datasets List Response}
\usage{
.parse_datasets_response(response)
}
\arguments{
\item{response}{The parsed response from on_request().}
}
\value{
A tibble with dataset information.
}
\description{
Converts the datasets() query response to a tibble.
}
\keyword{internal}
