% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-response.R
\name{.parse_files}
\alias{.parse_files}
\title{Parse Files Response}
\usage{
.parse_files(response)
}
\arguments{
\item{response}{The parsed response from on_request().}
}
\value{
A tibble with file information.
}
\description{
Converts the files query response to a tibble.
}
\keyword{internal}
