% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-response.R
\name{.parse_search_results}
\alias{.parse_search_results}
\title{Parse Search Results Response}
\usage{
.parse_search_results(response)
}
\arguments{
\item{response}{The parsed response from on_request().}
}
\value{
A tibble with dataset information.
}
\description{
Converts the search() query response to a tibble.
}
\keyword{internal}
