% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-response.R
\name{.parse_snapshots}
\alias{.parse_snapshots}
\title{Parse Snapshots Response}
\usage{
.parse_snapshots(response)
}
\arguments{
\item{response}{The parsed response from on_request().}
}
\value{
A tibble with snapshot information.
}
\description{
Converts the snapshots query response to a tibble.
}
\keyword{internal}
