% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-response.R
\name{.parse_subjects}
\alias{.parse_subjects}
\title{Parse Subjects Response}
\usage{
.parse_subjects(response, dataset_id)
}
\arguments{
\item{response}{The parsed response from on_request().}

\item{dataset_id}{The dataset identifier to include in output.}
}
\value{
A tibble with subject information.
}
\description{
Converts the subjects query response to a tibble.
}
\keyword{internal}
