% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-response.R
\name{.transform_timestamps}
\alias{.transform_timestamps}
\title{Transform Timestamps to POSIXct}
\usage{
.transform_timestamps(df)
}
\arguments{
\item{df}{A tibble with numeric timestamp columns.}
}
\value{
The tibble with POSIXct timestamp columns.
}
\description{
Converts numeric timestamp columns to POSIXct class.
}
\keyword{internal}
