% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject-filter.R
\name{is_regex}
\alias{is_regex}
\title{Check if Object is a Regex Pattern}
\usage{
is_regex(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"on_regex"}, \code{FALSE} otherwise.
}
\description{
Tests whether an object was created by \code{\link[=regex]{regex()}}.
}
\keyword{internal}
