% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openneuro-package.R
\docType{package}
\name{openneuro-package}
\alias{openneuro}
\alias{openneuro-package}
\title{openneuro: Programmatic Access to OpenNeuro Datasets}
\description{
Search, explore, and download datasets from 'OpenNeuro' \url{https://openneuro.org}, the largest open neuroimaging data repository. Queries the 'OpenNeuro' GraphQL API to discover datasets by modality, diagnosis, or keyword; inspect snapshots, files, and subject lists; and download full datasets or selected subsets via HTTPS, Amazon S3, or 'DataLad'. Downloaded data are cached locally so subsequent requests skip already-fetched files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbuchsbaum.github.io/openneuroR/}
  \item \url{https://github.com/bbuchsbaum/openneuroR}
  \item Report bugs at \url{https://github.com/bbuchsbaum/openneuroR/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com} [copyright holder]

}
\keyword{internal}
