% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{print.openneuro_handle}
\alias{print.openneuro_handle}
\title{Print Method for OpenNeuro Handle}
\usage{
\method{print}{openneuro_handle}(x, ...)
}
\arguments{
\item{x}{An \code{openneuro_handle} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
\code{x} invisibly.
}
\description{
Print Method for OpenNeuro Handle
}
