% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{transform_timestamps_posix}
\alias{transform_timestamps_posix}
\title{Transform Numeric Timestamps to POSIXct}
\usage{
transform_timestamps_posix(df)
}
\arguments{
\item{df}{A tibble with numeric timestamp column.}
}
\value{
The tibble with POSIXct timestamp columns.
}
\description{
Converts numeric timestamp columns to POSIXct class.
}
\keyword{internal}
