% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openneuro-package.R
\docType{package}
\name{openneuro-package}
\alias{openneuro}
\alias{openneuro-package}
\title{openneuro: Programmatic Access to OpenNeuro Datasets}
\description{
Search, explore, and download datasets from OpenNeuro, the largest open neuroimaging repository. Provides GraphQL-based discovery, multi-backend downloads (HTTPS/S3/DataLad), and smart caching.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbuchsbaum/openneuroR}
  \item Report bugs at \url{https://github.com/bbuchsbaum/openneuroR/issues}
}

}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
